# backend/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Phxntom Esports API",
    "description": "API backend pour la plateforme esports Phxntom. Utilisez ce fichier dans FlutterFlow pour importer tous les endpoints.",
    "version": "1.0.0",
    "contact": {
      "name": "Phxntom Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Développement local"
    },
    {
      "url": "https://votre-domaine.com/api",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/profile": {
      "get": {
        "tags": ["Profil"],
        "summary": "Récupérer son profil complet",
        "operationId": "getProfile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Profil détaillé",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Profil"],
        "summary": "Modifier son profil (pseudo, avatar)",
        "operationId": "updateProfile",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pseudo": { "type": "string" },
                  "avatar": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profil modifié",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
...
    "/accounts": {
      "get": {
        "tags": ["Comptes liés"],
        "summary": "Lister mes comptes de jeu liés",
        "operationId": "listAccounts",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Comptes liés",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string" },
                      "platform": { "type": "string" },
                      "gamertag": { "type": "string" },
                      "platformId": { "type": "string", "nullable": true },
                      "linkedAt": { "type": "string", "format": "date-time" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/accounts/link": {
      "post": {
        "tags": ["Comptes liés"],
        "summary": "Lier un compte manuellement (fallback OAuth)",
        "operationId": "linkAccount",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["platform", "gamertag"],
                "properties": {
                  "platform": {
                    "type": "string",
                    "enum": ["PSN", "XBOX", "STEAM", "BATTLE_NET", "ACTIVISION"]
                  },
                  "gamertag": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Compte lié"
          }
        }
      }
    },
    "/accounts/{id}": {
      "delete": {
        "tags": ["Comptes liés"],
        "summary": "Délier un compte",
        "operationId": "unlinkAccount",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Compte délié"
          }
        }
      }
    },
...
    "/loadouts": {
      "get": {
        "tags": ["Loadouts"],
        "summary": "Tous les loadouts (filtré ?weaponName=, ?category=)",
        "operationId": "listLoadouts",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "weaponName", "in": "query", "schema": { "type": "string" } },
          { "name": "category", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Loadouts"
          }
        }
      },
      "post": {
        "tags": ["Loadouts"],
        "summary": "Créer un loadout",
        "operationId": "createLoadout",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["weaponName", "category", "attachments"],
                "properties": {
                  "weaponName": { "type": "string" },
                  "category": { "type": "string" },
                  "attachments": { "type": "string" },
                  "description": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Loadout créé"
          }
        }
      }
    },
    "/loadouts/mine": {
      "get": {
        "tags": ["Loadouts"],
        "summary": "Mes loadouts personnels",
        "operationId": "myLoadouts",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Mes loadouts"
          }
        }
      }
    },
    "/loadouts/{id}": {
      "get": {
        "tags": ["Loadouts"],
        "summary": "Détail d'un loadout",
        "operationId": "getLoadout",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Loadout"
          }
        }
      },
      "delete": {
        "tags": ["Loadouts"],
        "summary": "Supprimer mon loadout",
        "operationId": "deleteLoadout",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Supprimé"
          }
        }
      }
    },
    "/loadouts/{id}/review": {
      "post": {
        "tags": ["Loadouts"],
        "summary": "Noter un loadout",
        "operationId": "reviewLoadout",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["rating"],
                "properties": {
                  "rating": { "type": "integer", "minimum": 1, "maximum": 5 },
                  "comment": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Avis ajouté"
          }
        }
      }
    },
    "/ai/queue/join": {
      "post": {
        "tags": ["Matchmaking IA"],
        "summary": "Rejoindre la file d'attente",
        "operationId": "joinQueue",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "gameMode": {
                    "type": "string",
                    "enum": ["KILLRACE", "RESURGENCE", "BATTLE_ROYALE"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dans la file",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "queueEntry": {
                      "type": "object",
                      "properties": {
                        "userId": { "type": "string" },
                        "pseudo": { "type": "string" },
                        "gameMode": { "type": "string" },
                        "rank": { "type": "string" },
                        "status": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ai/queue/leave": {
      "post": {
        "tags": ["Matchmaking IA"],
        "summary": "Quitter la file d'attente",
        "operationId": "leaveQueue",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "File quittée",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/ai/queue/count": {
      "get": {
        "tags": ["Matchmaking IA"],
        "summary": "Nombre de joueurs dans la file",
        "operationId": "queueCount",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Compteur",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ai/queue/process": {
      "post": {
        "tags": ["Matchmaking IA"],
        "summary": "Déclencher manuellement le matchmaking",
        "operationId": "processQueue",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "gameMode": {
                    "type": "string",
                    "enum": ["KILLRACE", "RESURGENCE", "BATTLE_ROYALE"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Matchmaking traité"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenu via /auth/login ou /auth/register"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "pseudo": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "avatar": { "type": "string", "nullable": true },
          "xp": { "type": "integer" },
          "rank": { "type": "string", "enum": ["BRONZE", "SILVER", "GOLD", "PLATINUM", "DIAMOND", "MASTER"] },
          "coins": { "type": "integer" },
          "wins": { "type": "integer" },
          "losses": { "type": "integer" },
          "totalMatches": { "type": "integer" },
          "trustScore": { "type": "integer" },
          "reportCount": { "type": "integer" },
          "dailyXp": { "type": "integer" },
          "dailyCoins": { "type": "integer" },
          "lastDailyReset": { "type": "string", "format": "date-time" },
          "isAdmin": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}